agent-browser: binary screenshot transport + extract panel modules
Transport: ship screenshot frames as raw bytes (Uint8Array over the webview
structured-clone channel) instead of base64. Removes the host-side
toString('base64'), the webview atob + per-byte decode loop, and ~33% transfer
inflation per shot — the largest, most frequent payload in the surface.
Dogfood-confirmed VS Code preserves the Uint8Array across the boundary.

Quality (pure extractions, behavior-preserving):
- use-pane-chrome.ts: the pane register/unregister + spawn-animation effects,
  previously copy-pasted byte-for-byte in TerminalPanel/IframePanel/
  AgentBrowserPanel, are now one shared hook.
- agent-browser-input.ts: the pure keyboard/mouse→CDP translation (VK tables,
  virtualKeyCode, modifiers, mouse maps, edit ops) lifted out of the panel, with
  agent-browser-input.test.ts covering the '.'=VK_DELETE footgun + modifier mask.
- agent-browser-screenshot-loop.ts: the backpressure state machine as
  createScreenshotLoop({getSession,getBinaryPath,isCapable,draw}); the panel
  wires it in a one-line useMemo.

AgentBrowserPanel drops ~190 lines and now reads as transport + sync + render.
