+++ lib/src/components/wall/IframePanel.tsx
@@ -1,11 +1,8 @@
-import { useContext, useEffect, useLayoutEffect, useMemo, useRef, useState } from 'react';
+import { useContext, useEffect, useMemo, useRef, useState } from 'react';
 import type { IDockviewPanelProps } from 'dockview-react';
 import { TERMINAL_BOTTOM_RADIUS_CLASS } from '../design';
-import {
-  FreshlySpawnedContext,
-  PaneElementsContext,
-  WallActionsContext,
-} from './wall-context';
+import { usePaneChrome } from './use-pane-chrome';
+import { WallActionsContext } from './wall-context';
 
 type IframePanelParams = {
   surfaceType?: string;
@@ -14,9 +11,8 @@ type IframePanelParams = {
 
 export function IframePanel({ api, params }: IDockviewPanelProps<IframePanelParams>) {
   const actions = useContext(WallActionsContext);
-  const { elements: paneElements, bumpVersion } = useContext(PaneElementsContext);
-  const freshlySpawned = useContext(FreshlySpawnedContext);
   const elRef = useRef<HTMLDivElement>(null);
+  usePaneChrome(api, elRef);
   const url = typeof params?.url === 'string' ? params.url : '';
   const origin = useMemo(() => {
     try {
@@ -41,37 +37,6 @@ export function IframePanel({ api, params }: IDockviewPanelProps<IframePanelPara
     const timer = setTimeout(() => setStalled(true), 5000);
     return () => clearTimeout(timer);
   }, [url]);
-
-  useEffect(() => {
-    if (!elRef.current) return;
-    paneElements.set(api.id, elRef.current);
-    bumpVersion();
-    return () => {
-      paneElements.delete(api.id);
-      bumpVersion();
-    };
-  }, [api.id, paneElements, bumpVersion]);
-
-  useLayoutEffect(() => {
-    const direction = freshlySpawned.get(api.id);
-    if (!direction) return;
-    freshlySpawned.delete(api.id);
-    const groupEl = api.group?.element;
-    if (!groupEl) return;
-    const className = `pane-spawning-from-${direction}`;
-    const animationName = `pane-spawn-from-${direction}`;
-    groupEl.classList.add(className);
-    const onEnd = (ev: AnimationEvent) => {
-      if (ev.animationName !== animationName) return;
-      groupEl.classList.remove(className);
-      groupEl.removeEventListener('animationend', onEnd);
-    };
-    groupEl.addEventListener('animationend', onEnd);
-    return () => {
-      groupEl.removeEventListener('animationend', onEnd);
-      groupEl.classList.remove(className);
-    };
-  }, [api, freshlySpawned]);
 
   return (
     <div
